+++ etlrules/__init__.py
@@ -7,11 +7,12 @@ from .exceptions import (
     InvalidPlanError, MissingColumnError, SchemaError, UnsupportedTypeError,
 )
 from .plan import Plan, PlanMode
+from .runner import load_plan, run_plan
 
 
 __author__ = """Ciprian Miclaus"""
 __email__ = "ciprianm@gmail.com"
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 
 
 __all__ = [
@@ -20,4 +21,5 @@ __all__ = [
     "ColumnAlreadyExistsError", "ExpressionSyntaxError", "GraphRuntimeError",
     "InvalidPlanError", "MissingColumnError", "SchemaError", "UnsupportedTypeError",
     "Plan", "PlanMode",
+    "load_plan", "run_plan",
 ]