+++ HISTORY.md
@@ -6,6 +6,8 @@
 * Apply substitution in the Read/Write rules for csv and parquet files for the file_name and file_dir parameters
 * Add a cli runner which allows users to run a yml file and parameterize with cli args the plan context
 * Add the csv2db plan/yml example
+* Add the db2csv plan/yml example
+* Remove poetry
 
 ## 0.2.2 (2023-11-26)